Extract a shared macro and add a few utilities
