+++ assets/sass/utilities/_utilities-display.scss
@@ -67,6 +67,10 @@
     padding-bottom: $spacingUnit * 1.5;
 }
 
+.u-no-padding {
+    padding: 0;
+}
+
 /* =========================================================================
     Margin
     ========================================================================= */
