+++ config/locales/cy.yml
@@ -19,7 +19,7 @@ global:
     about: Amdanom ni
     research: Ymchwil
     blog: Blog
-    updates: TODO
+    updates: Newyddion
   contact:
     city: Llundain
     phone: 0300 123 0735
