+++ config/locales/en.yml
@@ -20,7 +20,7 @@ global:
     research: Research
     about: About
     blog: Blog
-    updates: Updates
+    updates: News
   contact:
     city: London
     phone: 0345 4 10 20 30
