+++ controllers/blog/index.js
@@ -3,33 +3,9 @@ const path = require('path');
 const express = require('express');
 const { concat, get, isEmpty } = require('lodash');
 const { injectBreadcrumbs, injectBlogDetail, injectBlogPosts } = require('../../middleware/inject-content');
+const { buildPagination } = require('../../modules/pagination');
 
 const router = express.Router();
-
-/**
- * Build pagination
- * Translate content API pagination into an object for use in views
- */
-function buildPagination(paginationMeta) {
-    if (paginationMeta && paginationMeta.total_pages > 1) {
-        const currentPage = paginationMeta.current_page;
-        const totalPages = paginationMeta.total_pages;
-        const prevLink = `?page=${currentPage - 1}`;
-        const nextLink = `?page=${currentPage + 1}`;
-
-        return {
-            count: paginationMeta.count,
-            total: paginationMeta.total,
-            perPage: paginationMeta.per_page,
-            currentPage: currentPage,
-            totalPages: totalPages,
-            prevLink: currentPage > 1 ? prevLink : null,
-            nextLink: currentPage < totalPages ? nextLink : null
-        };
-    } else {
-        return;
-    }
-}
 
 function renderListing({ res, title, entries = [], entriesMeta = null, breadcrumbs = [] }) {
     const pagination = buildPagination(entriesMeta.pagination);
