+++ modules/pagination.js
@@ -1,0 +1,28 @@
+'use strict';
+
+/**
+ * Build pagination
+ * Translate content API pagination into an object for use in views
+ */
+function buildPagination(paginationMeta) {
+    if (paginationMeta && paginationMeta.total_pages > 1) {
+        const currentPage = paginationMeta.current_page;
+        const totalPages = paginationMeta.total_pages;
+        const prevLink = `?page=${currentPage - 1}`;
+        const nextLink = `?page=${currentPage + 1}`;
+
+        return {
+            count: paginationMeta.count,
+            total: paginationMeta.total,
+            perPage: paginationMeta.per_page,
+            currentPage: currentPage,
+            totalPages: totalPages,
+            prevLink: currentPage > 1 ? prevLink : null,
+            nextLink: currentPage < totalPages ? nextLink : null
+        };
+    }
+}
+
+module.exports = {
+    buildPagination
+};
